refactor: move fixture to main conftest
